feat: add /send handler

+++ internal/models/message.go
@@ -1,8 +1,8 @@
 package models
 
 type Message struct {
-	FromUserID uint `gorm:"not null;index"`
-	ToUserID   uint `gorm:"not null;index"`
+	FromUserID int64 `gorm:"not null;index"`
+	ToUserID   int64 `gorm:"not null;index"`
 	Text       string
 	FromUser   User `gorm:"foreignKey:FromUserID;references:ID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;"`
 	ToUser     User `gorm:"foreignKey:ToUserID;references:ID;constraint:OnUpdate:CASCADE,OnDelete:CASCADE;"`
